Reorg widget extension files
